requirements.txt
beautifulsoup4
feedgen
python-dotenv
requests
